+++ lib/japanese_address_parser.rb
@@ -5,13 +5,24 @@ require_relative 'japanese_address_parser/address_parser'
 require_relative 'japanese_address_parser/version'
 
 module JapaneseAddressParser
-  module_function
-
   def call(full_address)
+    _call(full_address)
+  rescue ::JapaneseAddressParser::NormalizeError
+    nil
+  end
+
+  def call!(full_address)
+    _call(full_address)
+  end
+
+  def _call(full_address)
     normalized = ::JapaneseAddressParser::AddressNormalizer.call(full_address)
 
     # このライブラリで探索するのは町域まで。
     # それ以降のデータを使って探索するとデータと名前が一致しないことがあるので、町域までのデータを使う。
     ::JapaneseAddressParser::AddressParser.call("#{normalized['pref']}#{normalized['city']}#{normalized['town']}")
   end
+
+  module_function :call, :call!, :_call
+  private_class_method :_call
 end