+++ CHANGELOG.md
@@ -9,9 +9,6 @@ Change Log の形式は [Keep a Changelog](http://keepachangelog.com/) に従い
 
 ### Changed
 
-- [#45](https://github.com/yamat47/japanese_address_parser/pull/45) geolonia/japanese-addressesのバージョンを上げた。([@yamat47](https://github.com/yamat47))
-- [#46](https://github.com/yamat47/japanese_address_parser/pull/46) CIでRSpecを並列実行させるようにした。([@yamat47](https://github.com/yamat47))
-
 ### Deprecated
 
 ### Removed
@@ -19,6 +16,12 @@ Change Log の形式は [Keep a Changelog](http://keepachangelog.com/) に従い
 ### Fixed
 
 ### Security
+
+## [2.2.1] - 2022-07-16
+### Changed
+
+- [#45](https://github.com/yamat47/japanese_address_parser/pull/45) geolonia/japanese-addressesのバージョンを上げた。([@yamat47](https://github.com/yamat47))
+- [#46](https://github.com/yamat47/japanese_address_parser/pull/46) CIでRSpecを並列実行させるようにした。([@yamat47](https://github.com/yamat47))
 
 ## [2.2.0] - 2022-03-12
 ### Added